% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_country.R
\name{nd_available_countries}
\alias{nd_available_countries}
\title{Get available countries}
\usage{
nd_available_countries()
}
\description{
Get available countries
}
\examples{
 library(nagerdate)

 nd_available_countries()

}
