% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_country.R
\name{nd_country_info}
\alias{nd_country_info}
\title{Get country info}
\usage{
nd_country_info(country_code)
}
\arguments{
\item{country_code}{Country code.}
}
\description{
Get country info
}
\examples{
 library(tidyverse)
 library(nagerdate)

 de <- nd_country_info("DE")

 de_borders <- de |> tidyr::unnest(borders, names_sep = "_")

}
