% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_holiday.R
\name{nd_holidays}
\alias{nd_holidays}
\title{Get country info}
\usage{
nd_holidays(country_code, year)
}
\arguments{
\item{country_code}{Country code.}

\item{year}{Year.}
}
\description{
Get country info
}
\examples{
 library(nagerdate)

 de <- nd_holidays("DE")


}
